Set up InterruptableJob
We need to be able to cancel backgruond compilation if the hybrid
backend managed to finish the job through the interpreter.
This was currently hacked together and we needed to sleep long enough to
make sure a detached background thread would finish compilation.

We now have a proper interrupt handler called InterruptableJob which
allows us to cancel background compilation properly. Internally, this
uses process and event file descriptors to implement efficient
interruption.