Start working on filters
Filters are the first inkfuse operator which changes output cardinality
and thus rescopes a pipeline.
This commit adds foundational work for filters. It adds the required
operators and sub-operators.

It also fixes the IU logic propagating through trees. We now use
unordered_sets rather than vectors to be more robust towards multiple
suboperators reusing the same IU.
This makes it a bit harder to implement suboperators which require IU
ordering in their operands. We might have to change the abstraction
again down the line, but for now it's fine.