Rework the IU Scoping Model
This commit rebuilds the IU scoping to make it easier to understand. Our
previous approach to IU scoping was (sadly) broken.

We now define a more principled approach. A scope contains a set of IUs
mapping to a globally unique variable identifier, and to a producing
operator. A scope has a unique selection vector.

If an operator rescopes a pipeline, it build a new scope which either
lets certain previous IDs through, or clears them. For example, a filter
will let through previous IDs. However, the filter will define a new
selection IU.
A join will clear the IDs, as it needs to redefine new vectors for the
respective IUs.

Within a CompilationContext, the operators do not need to resolve their
own scope anymore. Rather, they pass themselves to the
CompilationContext which will perform scope resolution and provide the
correct declarations.