Move compiler to clang-11 (#8)
Currently, the repipe on filters does not work in gcc-9. I tried
debugging this for some time but to no avail.

The problem is that no FuseChunkSink gets added to the pipeline 
for some reason. As such, there is no output operator. Something goes
wrong in the std::for_each on materialite_all. 

As things work flawlessly across clang versions, let's stick with clang
in the CI for now.