Add Interpreted Filters (#9)
We now can interpret filters. These simply generate an empty fragment.
We currently don't support reading over filtered chunks, we have to add
this to the loop drivers.

In a first step however, we want to evaluate if our current filter model
even makes sense.

CI is failing because (I think) we don't have clean argument ordering on 
the source IUs at the moment. In the release build, the arguments on the
subtraction get mixed up and we compare the wrong columns. This is 
a pretty glaring issue I will have to fix in a follupw PR to not make this 
one too crowded.