[Draft] Drop IU Scopes (#11)
This commit moves away from Scoped IUs to a flat model.
Through this, bookkeeping becomes significantly simpler.

As part of this, we simplify the Operator and Suboperator
interfaces considerably. We can now always talk about raw IUs
and never have to worry about the previously brittle scoping.

We also introduce the concepts of strong links in the suboperator
DAG. Strong links allow us to express that a chain of suboperators
should be interpreted as one, rather than individually.