Add runtime for Hashing
This commit adds runtime support for hashing.
We use xxhash as our default hash function for now.

As part of this, we add the rdynamic flag in order to allow the compiled
code which gets dlopened to access the runtime symbols of inkfuse.
This is not an incredibly clean approach, but will work well enough for
now.