Start work on LazyParameter
As we are working towards aggregations, we need support for lazy
parameters that can either be hard-baked into the code or read from the
global state during runtime.

This is needed to create primitives like addition with a constant.
Within the generated code, we want to hard-bake the value into the
compiled query binary.
At the same time, we cannot generate a query for every possible value
that can be a constant addition in SQL, as there are infinitely many.

For aggregations specifically, the lazy parameters will be needed for
offsetting within packed keys and values of the aggregation hash-tables.