Refactor & Clean Up
This commit does some cleanup to make it easier to implement
aggregations. We perform the following two things:

We rename LAZY_PARAM to RUNTIME_PARAM as it's more descriptive and lazy
is a really loaded term. I thought it would be fitting but it isn't
really.

We remove the current concept of runtime parameters. These are currently
objects we can attach to operators like table scans that are needed to
set up the runtime global state of the operator, but not needed for code
generation.
While this is in principle a valuable distinction to make, it added too
much complexity and also wasn't built perfectly.
The new approach is much simpler - add the old runtime params to the
constructor of the object and substitute default values if none are
available.
This requires a bit more attention when instantiating these objects
in the context of a specific query. At the same time, it reduces
complexity of most operators as these don't require differential
treatment.