Set up HashTableRuntime
This commit introduces a working version of a HashTableRuntime.
The HashTableRuntime provides access to a backing hash table that can be
used within aggregations and hash joins.

Currently, it is an extremely simple power-of-two hash table that grows
by 2x once the fill-factor grows beyond 50%.

We can significantly improve that hash table in the future, but it
serves us decently well for a v0.