Change HtLookupSubop to RuntimeFunctionSubop
We will need runtime function lookups in more places. We don't want a
single suboperator for every runtime function.

This commit turns the HtLookupSubop into a more general purpose
RuntimeFunctionSubop.

The RuntimeFunctionSubop allows calling an arbitrary runtime function on
a runtime object. It gets a function name, an IU as parameter and an object in the
runtime state. In the compiled code it invokes the function name with
the runtime object as the first parameter and the IU as the second
parameter.