Build Key-Aware Hash Tables
As it turns out, the current approach to hash tables within InkFuse did
not really work for aggregations.
This is because collision resolution through chained suboperators turned
out to be a lot harder than initially thought.

The core problem we ran into was that within a chunk for
collision-resolution, the same key could be stored multiple times. In
this case, we would have to make sure that only the first key would get
inserted into the hash table once.

We decided to instead implement a key aware hash table. This hash table
is able to do collision resolution by itself. For now we only support
simple compound keys which can be memcmpared, but the idea can be
extended to hash tables with string support as well.

The hash tables are general enough to support more complex operations in
the future:
- Collations: We can extend the hash tables with complex keys to have
collation-based comparators. The complex key comparison would then be an
array of function pointers, one pointer for every slot.
- Inequality joins/joins with complex key conditions: For complex joins,
the matching of join rows is usually done when iterating the chain of
values. We still have some "equality-part" which is turned into the key
condition of the hash join, more complex predicate evaluation happens
afterwards when traversing the rows for a specific equality match.

This commit also introduces a hash-table microbenchmark where we compare
ourselves to std::unordered_map. We outperform std::unordered_map across
the board.