Add ByteArray Type
This commit introduces a new ByteArray type. The ByteArray can be used
to efficiently represent packed keys.
This is needed for the aggregation pipeline, as we first pack the key
and then pass a pointer to the packed key to the aggregation hash table.

The main challenge when building a ByteArray type is that the static
width of the array needs to be erased from the vectorized primitives. We
cannot generate primitives for all possible key widths, as these are
(almost) infinite.

We thus define the width of the ByteArray as runtime parameter of the
IndexedIUProvider. An IndexedIUProvider which goes over a parametrized
type carries its own efficient iterator to provide efficient access to
the packed keys.

We will introduce tests in a followup commit.