Add Initial Version of ScratchPadIUProvider
We need a dedicated source-style operator to allow us to work with packed keys
in the context of a compiled query.

In interpreted mode, the fixed-size byte array is simply added to the FuseChunks,
this is elegant as then allocation size decisions can be moved to query runtime
where the size of a ByteArray is known.

This however does not really work for compiled execution, where we don't read the
ByteArrays from storage. Rather, the ByteArray is a transient scrach-space needed
to e.g. pack keys for later insertion into a hash table.

The ScratchPadIUProvider allows us to do this. It is only evaluated for compiled
queries and allocates the required "scratch space" within the function preamble.
This way, many tuples from the source relation can share the same scratch space,
which should lead to improved performance.

As part of these changes, we also add a first version of dynamic memory allocation
to a Pipeline. The MemoryRuntime allows easy allocation of "scratchpad memory" which
gets dropped as soon as a pipeline is finished.