Get ScratchPadIUProviders to Work
This commit introduces a new, succeeding unit tests for the ScratchPadIUProviders.
Getting them working required larger infrastructure changes.

We now support pseudo IUs for the packing operators. This is needed to ensure
that during fused code generation, the packing logic gets generated before
the actual hash table insert.

Before, this was not possible. It was non-determinisitc what code would be generated
first.

Arguably, the pseudo-IUs are not a perfect abstraction. There are two
types of dependencies in an incremental fusion graph:
- IU dependencies, i.e. when a suboperator reads an IU that has to be provided first
- Logic dependencies, i.e. when the code of a suboperator can only be genererated
  after the code of another suboperator was generated.

Any IU dependency implies a logic depedency. However, not every logic dependency
implies an IU depedency.
In InkFuse, the logic dependencies that are not IU dependencies are modeled through
pseudo-IUs with void type.

This leads to some weird handling of corner cases that might be confusing for newcomers.
For example, the repiping operation that inserts scans from a FuseChunk cannot just iterate
over all input IUs. Rather, it needs to ignore void-typed IUs that model logic dependencies
that are not an IU dependency.