Set up HashTableSource
This commit introuces a new HashTableSource. The HashTableSource allows
us to scan data from a hash  table within the generated/vectorized code.

This is needed for operators that read from a hash table.
Examples are reading from an aggregate hash table or a join hash table.
The HashTableSource simply produces a char*.

The HashTableSource stores an iterator into the hash table within its global
state. Through a runtime call, this iterator can then be advanced.
We make sure to choose morsels that have at most the vectorized chunk size.

The next steps in finalizing the initial aggreagtion support are:
- Hooking up the HashTableSource into KeyUnpacking suboperators. We need
  properly vectorized fragments for that.
- Creating and end-to-end test of the aggregaetion pipeline.