Get First Aggregation Working Correctly
This commit fixes the remaining logical correctness issues we had
in our aggregation. Apart from some minor offset calculation bugs,
the largest issue was more fundamental:

We have a single primitive that is responsible for calling the runtime
function ht_sk_lookup_or_insert. Within a primitive, this iterates
over a chunk of input IUs and tries to lookup those IUs in the hash table,
or insert them if they did not exist.

This caused a subtle problem during vetoritzed interpretation:
What if the hash table gets resized in the middle of the current chunk?
Here, we could run into cases where the pointers before the resize would
now point to deallocated memory from the smaller hash table before the
resize.

This is an artifact of us using a linear probing hash table where pointers
to the aggregation state are unstable.

We now resolve this problem by allowing a growing hash table to set a
flag within the query context indicating that the current morsel has to
be repeated. In this case we flush the invalid state and rerun the morsel.
This is possible because the primitive is idempotent.

An alternative would have been to move to a chaining hash table, but this is
still an interesting concept to retain.