Introduce Basic String and Date Types
This commint introduces basic string and date type support in inkfuse.
Internally, dates are represented as day offset to the epoch 1-1-1970.
Strings are just represented as char*.

In terms of function support, the commit introduces the following:
- It adds the required runtime functions for date types
- It adds strcmp for string types, this is enough for most of TPC-H
These are already wired up into the primitives for vectorized code
generation.

A meta-point: string handling in engines usually is pretty complex.
The main challenge is ownership: when do you allocate/deallocate strings
that are e.g. created by some regex function?
In inkfuse this is much simpler for now as we don't have functions
mutating strings yet. As such, ownership is completely hidden in the
storage layer and we just pass strings around.