Allow Ingesting Data From Text
This commit allows loading data into columns from text.
For all the TPC-H types that we need, we now allow deserializing data from
a text-based representation into the actual PODColumn/StringColumn.

For this, we turned the previous TypedColumn into a PODColumn that
directly has an inkfuse type attached.
Before, the TypedColumn was templated over a C++ storage type. This was
not clean as different InkFuse types can have the same storage type. The
easiest examples are dates which are internally stored as int32_ts.

By making the PODColumn dependent on the InkFuse type, we can now have
different deserialization logic for the same underlying storage type.
This is tested in `test/test_storage.cpp`.