Set Up q1 Scaffold
This commit sets up the scaffolding needed to
run TPC-H q1. We add a simplified version of the q1 query plan
and run it in the test_queries tests.

The simplified plan runs through on all three execution modes.
It still has the following deviations from regular q1 that will
be fixed later:
- Only computing one aggregate function
- Grouping by only one column
- No order by