Add Avg and Sum Aggregation
This commit introduces Avg and Sum aggregate functions in InkFuse. We
need these to be able to run TPC-H Q1.

Avg is a very interesting aggregate function as it's decomposed into two
separate state granules. It utilizes a count and a sum within the
aggregation state in order to compute the final avg result.

This is neat, as this means the avg can share aggregate state with other
operators.