Fix Intermediate Column Sizes
Our intermediate column sizing in the face of scratch pad IUs was
broken during interpreted execution.

As these scratch pad IUs are never written to in an expression that has
the scratch pad IU as a sink, their size was also never set up properly.
For an aggregation pipeline like TPC-H Q1 this meant that the scattering
of compound aggregation keys worked fine, but then things took a turn
for the worse.

As the size of the compound key column was always 0, no remaining work
would happen.

We now mitigate this with a nasty hack in the Executor. If we detect
that we are writing into a scratch pad IU, we force the size to be the
current morsel size. This should be designed more cleanly in the future.