Overlapping Pipeline Compilation
This commit makes our pipeline compilation overlapping.
So far the query executor ran one pipeline after the other,
only triggering compilation when a pipeline is ready to be executed.

This means that for compiled mode we would stall #pipelines waiting on
compiled code to be ready.
It's much smarter to trigger compilation for all queries first, and then
start executing one after the other. There's a good chance that only the
first pipeline is blocked on compilation in this case.

This e.g. makes compiled TPC-H Q1 on small scale factors 2x faster, because we
only have to wait on compilation once, not twice.