New InkFuse Runner
The old runner was still an artifact of us only supporting expressions.
Now that we support real queries, it's time to get much more cool.

This commit reimplements our runner - it can now load data at different
scale factors and run Q1 on top of it. With the nice output formats and
timing information.