Add Primary Key Joins
Primary key joins are a simple version of a join. They are simple
because we don't have to worry about duplicate keys in a hash table, and
we also don't have to worry about growing chunks.

This commit implements this type of join variation. This enables us to
add the first TPC-H join queries, as these join types are common in
TPC-H.

One thing we didn't think of yet that this commit fixes is what happens
with suboperators that have the same IU as input multiple times. This
can e.g. happen in a join probe when filtering on the pointer result of
a lookup, and also filtering those pointers into a dense column of
matches.

We now generate separate fragments for filters on these cases. This
approach could be extended to expressions as well.