Make InkFuse reproducible
This commit makes the benchmarking of InkFuse reproducible.
It generates a new `inkfuse_bench` binary, generated TPC-H data, and
invokes the benchmarking binary.

The benchmarking binary generates the correct result csv files.

The commit also contains improvements to the performance of our Hybrid
backend. We do not wait for the compiling background thread anymore and
thus subprocess termination.
Rather, we detach the thread which cleans up asynchronously.