Fix Performance Problems with Hybrid Backend
We didn't design the IR and backends carefully. They abuse a ton of
shared pointers, don't generate a compact IR, and have a bunch of
performance problems.

Since we need the IUs from the query root to generate C, our C
generation so far happened within the hot path of the hybrid backend.

This lead to multiple milliseconds of overhead when running queries on
this backend.
This commit gets the performance of the hybrid backend much closer to
the performance of the interpreted backend.

We do this by performing C generation in an asynchronous fashion on a
background thread.
This is done by adding a control block to the query execution. This
control block is passed to the background threads and keeps the
relational algebra tree alive.

If we built the system from scratch today we should pay much more
attention to IR design.