Performance Improvements
This commit contains a variety of performance improvements to InkFuse.
These allow us to get closer to the performance of Umbra. Specifically,
we do the following:

- fork -> vfork. So far we used subprocess.h to perform background
compilation of C code. This caused significant performance loss in the
hybrid backend. The reason is that during the fork, the page tables of
the parent are set readonly. Every page fault then causes copy on write.
We now don't fork anymore but use vfork. The forked child instantly
calls execv. This improves performance significantly on the hybrid
backend, finally allowing it to perform on-par with the interpreter.

- Hybrid hyrbird backend. When we see that the generated code is slower
than the vectorized interpreter, we stick with the vectorized
interpreter.