Add Compilation Time to Reproduction Scripts
This commit adds the time spent waiting for query compilation to be
finished to the reproduction scripts.

The inkfuse_bench target now reports the duration that the compiled
backend was spent on code generation. For DuckDB, this is always set to
zero as no code has to be generated.

The duration is added to the evaluation plots, represented by the shaded
area of each bar. This makes it easier to see where time is spent.