Add Perf Microbenchmarks
This commit introduces microbenchmarking support for the inkfuse_runner,
as well as scripts in `reproducibility` that make it easy to analyze the
metrics.

We now have an optional argument on the inkfuse_bench binary that
performs the experiments in microbenchmarking mode.
It measures perf counters and dumps the results into a csv file.
Note that in microbenchmarking mode, no query latencies are written as
we need some explicit sleep calls between comiling & execution to
properly perf the runtime behaviour. Because of this the timing results
on the compiled backend become unreliable.