Prepare Pipeline Executor for Multithreading
This commit prepares the PipelineExecutor for multithreading. We can now
spawn multiple threads that execute morsels at the same time.

In the PipelineExecutor, the new control flow when calling `runPipeline` is:
1. Single threaded preparation work - compiling, setting up the
   interpreters, etc
2. Run all morsel swimlanes in parallel. One morsel swimlane takes
   morsels from the source through the entire pipeline until the source
   is exhausted.
   This phase spawns worker threads if we need more than one thread.
   The thread that `runPipeline` was called on always runs on morsel lane 0.
3. Wait for all morsel swimlanes to be finished, then the query is done.

There is one thing we need to pay attention to here: setting up the
compiled runner state if we are in hybrid mode. We need to make sure
that this is only done from one thread and all other threads are
waiting. For this we introduce a new `OnceBarrier` that is a mix of
std::once and std::barrier with the special property that the _last_
thread registering with the barrier calls the once.