Make MemoryRegions ThreadLocal
We need the `restart_flag` for morsel restarts in the interpreter to be
thread-local, as well as the memory allocators used to e.g. create
scratch-pad IUs.

At the moment inkfuse doesn't allocate a lot of memory from within the
generated code, but it writes to the allocated memory _a lot_. Luckily
the allocated regions always span multiple full cache lines, so we can
be sure that the scratch pad IUs are exclusive to the threads working on
them.

If we ever allocate more, the memory regions are 64 byte aligned,
allowing threads to allocate without the cache lines containing the
arena metadata to bounce.