Get First Multithreaded Test to Run
We now have multithreaded plans looking like:
Scan -> Expression -> Filter -> Print + Count
Running in InkFuse, with tests validating that the results are correct.

This means we have successfully hooked up all the basic infrastructure
for multithreaded query execution using morsel-driven parallelism.

All that's left (as if it was that easy ;-)) now is to make the
aggregations and joins multithreaded.