Deferred Runtime State Initialization
At the moment we always set up the runtime state during the construction
of the initial PipelineDAG.
This was never a clean design, but now it's coming to bite us. When we
construct the PipelineDAG we don't know the thread count yet. As a
result we also can't set up the MaterializationBuffers for all worker
threads.

This leads to a lot of problems with how we set up operator state. This
commit cleans up these abstractions a bit. During the setup of the
PipelineDAG we now only create a `DeferredState`, which is initialized
once we set up the state of the suboperators.

This allows us to actually set up different TupleMaterializers for the
different threads in the join.