Remove restrict pointers from the fragments for now
These make it much harder to actually read the fragments, and when I
benchmarked it it made almost no difference. Will re-measure eventually
and see if we should add them again.

The problem is that while these work nicely for the toy engine, it's
harder to make these guarantees in a real engine. For example:

In `SELECT a + a FROM T`, we most likely consume the same column pointer
twice. Making these pointers restrict is dangerous.