Allow Supberators to Only Generate Code for the Vectorized Backend
This commit is the next one in the chain to generate a faster vectorized
backend. We can now make both a `Suboperator` and the
`CompilationContext` with additional optimziation hints.

This allows us to mark suboperator that generate prefetching code in a
way that does not generate code for operator-fusing codegen.
The prefetching calls are now only emitted for functions in the
vectorized backend, but do not generate code for compiled execution.

In general, the prefetching for operator fusing code is not important,
as we will do a lookup on the same tuple right after, which will then
cause the respective cache miss. As a result, prefetching only generates
more instructions and  function calls.