Interpreter Performance Tuning
If we want to effectively use prefetching for our hash tables we need to
make sure that we don't evict some of the prefetched data because our
chunks are too large.

Our microbenchmarks show that this means we can go up to a maximum chunk
size of about 512 before hash table throughput starts dropping again.

This commit reduces the morsel size to 512 and improves performance in
the interpreter backend sufficiently that we are able to retain the same
performance.

We were doing an excessive number of dynamic casts in the interpreted
driver that started showing up in our perf traces once we dropped to a
chunk size of 512 or 256 rows per morsel.
We now only do those dynamic casts once during interpreter setup and
then cache the appropriately casted objects. This is especially
important for the special casing of the FuseChunks.