Set CPU Affinities for Thread Workers
This doesn't have a significant impact on overall system performance,
but it's still the right thing to do.

We now ensure that worker threads are pinned to a CPU core during query
processing. Worker 0 is pinned to CPU 0, worker 1 is pinned to CPU 1,
and so on.

This makes sure that we don't bounce our caches so much when a thread
gets rescheduled during morsel execution. I had hoped this would
positively impact ROF performance, but it really doesn't.